% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_signal}
\alias{filter_signal}
\title{Filter a signal}
\usage{
filter_signal(
  signal,
  filter = "sgolay",
  order = 2,
  window_length = NULL,
  cutoff_freq = NULL,
  sampling_freq = NULL,
  type = "low",
  apply = 1
)
}
\arguments{
\item{signal}{Signal to filter.}

\item{filter}{Type of filter (default is \code{"sgolay"}, or \code{"butter"}).}

\item{order}{Order of the filter.}

\item{window_length}{Window length of the Savitzky-Golay filter.}

\item{cutoff_freq}{Cut-off frequency of the Butterworth filter.}

\item{sampling_freq}{Sampling frequency of the signal.}

\item{type}{Butterworth band type (default is \code{"low"}).}
}
\value{
The filtered signal.
}
\description{
Filter a signal
}
