% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_polar_ci}
\alias{geom_polar_ci}
\title{Polar confidence intervals.}
\usage{
geom_polar_ci(data, group = NULL, ci_z = 1.96, ci_alpha = 0.1)
}
\arguments{
\item{data}{A tibble which is the output of \link[rticulate]{predict_polar_gam} with the argument \code{return_ci = true}.}

\item{group}{The optional grouping factor.}

\item{ci_z}{The z-value for calculating the CIs (the default is \code{1.96} for 95 percent CI).}

\item{ci_alpha}{Transparency value of CIs (the default is \code{0.1}).}
}
\description{
It provides a `geom` for plotting polar confidence intervals from the output of \link[rticulate]{predict_polar_gam} with the argument \code{return_ci = true}.
}
