% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{get_acceleration}
\alias{get_acceleration}
\title{Get acceleration of displacement}
\usage{
get_acceleration(signal)
}
\arguments{
\item{signal}{The signal to get the acceleration of.}
}
\value{
A vector with the second derivative of the signal.
}
\description{
Get acceleration of displacement
}
