% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{get_velocity}
\alias{get_velocity}
\title{Get velocity of displacement}
\usage{
get_velocity(signal)
}
\arguments{
\item{signal}{The signal to get the velocity of.}
}
\value{
A vector with the first derivative of the signal.
}
\description{
Get velocity of displacement
}
