% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palate}
\alias{palate}
\title{Palate profile dataset.}
\format{
A data frame with 42 rows and 14 variables.
\describe{
    \item{speaker}{speaker ID}
    \item{seconds}{time of coordinate, in seconds}
    \item{rec_date}{date and time of recording}
    \item{prompt}{prompt string}
    \item{label}{label of annotation}
    \item{TT_displacement}{smoothed displacement of tongue tip}
    \item{TT_velocity}{velocity of tongue tip displacement}
    \item{TT_abs_velocity}{absolute velocity of tongue tip displacement}
    \item{TD_displacement}{smoothed displacement of tongue dorsum}
    \item{TD_velocity}{velocity of tongue dorsum displacement}
    \item{TD_abs_velocity}{absolute velocity of tongue dorsum displacement}
    \item{fan_line}{fan line number}
    \item{X}{horizontal coordinate at time \code{seconds}}
    \item{Y}{vertical coordinate at time \code{seconds}}
}
}
\usage{
palate
}
\description{
A dataset containing the palate profile of a single speaker.
}
\keyword{datasets}
