% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_polar_smooths-deprecated}
\alias{plot_polar_smooths-deprecated}
\title{Plot smooths from a polar \code{gam}}
\arguments{
\item{split}{Columns to separate as a named list.}

\item{sep}{Separator between columns (default is \code{"\\."}, which is the default with \code{}). If character, it is interpreted as a regular expression.}
}
\value{
An object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
It plots the smooths of a polar GAM fitted with \code{polar_gam()}.
}
\seealso{
\code{\link{rticulate-deprecated}}
}
\keyword{internal}
