% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticulate.R
\name{plot_splines}
\alias{plot_splines}
\title{Plot tongue contours from spline data.}
\usage{
plot_splines(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
It plots tongue contours from data imported from AAA.
}
