% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tongue}
\alias{plot_tongue}
\title{Plot tongue contours from spline data.}
\usage{
plot_tongue(data, geom = "line", ..., palate = NULL, palate_col = "green")
}
\arguments{
\item{data}{A data frame with splines data.}

\item{geom}{Type of geom to plot. Possible values are: \code{line} (the default),
\code{point}, \code{path}.}

\item{...}{List of arguments to be passed to \code{geom}.}

\item{palate}{An optional data frame with the palate spline. If provided,
the palate is plotted.}

\item{palate_col}{The colour of the palate spline (the default is \code{green}).}
}
\description{
It plots tongue contours from data imported from AAA.
}
