% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{read_ag500_pos}
\alias{read_ag500_pos}
\title{Read EMA data from AG500 pos files}
\usage{
read_ag500_pos(path, channels = 12, ch_values = 7, bytes = 4, fs = 200)
}
\arguments{
\item{path}{Path to the .pos file.}

\item{channels}{Number of channels (default \code{12}).}

\item{ch_values}{Number of values per channel (default \code{7}).}

\item{bytes}{Number of bytes per value (default \code{4}).}

\item{fs}{Sampling frequency (default \code{200} Hz).}
}
\value{
A tibble.
}
\description{
Read EMA data from AG500 pos files
}
