% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{resample_signal}
\alias{resample_signal}
\title{Resample (up or down) a signal}
\usage{
resample_signal(signal, time, by = 2, to, from, method = "interpolation")
}
\arguments{
\item{signal}{The signal to resample.}

\item{time}{The time vector of the signal to resample.}

\item{by}{The factor by which to resample the signal (default is \code{2}).}

\item{to}{The frequency to resample to.}

\item{from}{The original sampling frequency.}

\item{method}{Resampling method (default is \code{interpolate} which uses \code{approx}).}
}
\value{
A list with the resampled signal and time if \code{methdo = "interpolate"}.
}
\description{
Resample (up or down) a signal
}
