% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticulate.R
\docType{package}
\name{rticulate}
\alias{rticulate}
\alias{rticulate-package}
\title{rticulate: A small package for Ultrasound Tongue Imaging in R.}
\description{
This is a package providing a function for processing AAA splines
files and plot tongue contours.
}
