% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stimuli}
\alias{stimuli}
\title{Stimuli dataset.}
\format{
A data frame with 12 rows and 11 variables.
\describe{
    \item{item}{item ID}
    \item{word}{words of the form CVCV}
    \item{ipa}{IPA transcription of the words}
    \item{c1}{first consonant}
    \item{c1_phonation}{phonation of the first consonant, \code{voiceless}}
    \item{vowel}{first and second vowel}
    \item{anteropost}{backness of the vowel, \code{back} or \code{central}}
    \item{height}{height of the vowel, \code{high}, \code{mid} or \code{low}}
    \item{c2}{second consonant}
    \item{c2_phonation}{phonation of the second consonant, \code{voiceless} or \code{voiced}}
    \item{c2_place}{place of the second consonant, \code{coronal} or \code{velar}}
}
}
\usage{
stimuli
}
\description{
A dataset with linguistic information on the stimuli.
}
\keyword{datasets}
