% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_ci}
\alias{transform_ci}
\title{Transform confidence intervals to cartesian}
\usage{
transform_ci(predictions, origin = NULL)
}
\arguments{
\item{predictions}{A data frame containing the predictions obtained from a
polar \code{gam}/\code{bam}.}

\item{origin}{The coordinates of the origin as a vector of \code{c(x, y)}
coordinates.}
}
\description{
It returns the cartesian coordinates of the confidence intervals from a polar
\code{gam}/\code{bam}.
}
\keyword{internal}
