% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinematics.R
\name{get_landmarks}
\alias{get_landmarks}
\title{Get gestural landmarks}
\usage{
get_landmarks(signal, time, threshold = 0.2)
}
\arguments{
\item{signal}{The displacement signal.}

\item{time}{The time of the signal.}

\item{threshold}{The velocity threshold (default is \code{0.2}, corresponding to 20 percent velocity.)}
}
\value{
A tibble with one row and a column for each gestural landmark.
}
\description{
Get gestural landmarks
}
