% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{get_origin}
\alias{get_origin}
\title{Get the origin of spline data}
\usage{
get_origin(data, fan_lines = c(10, 25))
}
\arguments{
\item{data}{The spline data (the cartesian coordinates must be in two columns named \code{X} and \code{Y}).}

\item{fan_lines}{A numberic vector with two fan lines (the default is \code{c(10, 25)}).}
}
\description{
It returns the x,y coordinates of the intersection of the fan lines, which correponds to the origin of the ultrasonic waves/probe surface.
}
