% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_polar_smooths}
\alias{plot_polar_smooths}
\title{Plot smooths from a polar \code{gam}}
\usage{
plot_polar_smooths(model, time_series, comparison, origin = NULL,
  facet_terms = NULL, conditions = NULL, exclude_random = TRUE,
  series_length = 100)
}
\description{
It plots the smooths of a polar GAM fitted with \code{polar_gam()}.
}
