% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_splines}
\alias{plot_splines}
\title{Plot tongue contours from spline data.}
\usage{
plot_splines(data, ..., palate = NULL, palate.col = "green")
}
\arguments{
\item{data}{A data frame with splines data.}

\item{...}{List of arguments to be passed to \code{geom}.}

\item{palate}{An optional data frame with the palate spline. If provided,
the palate is plotted.}

\item{palate.col}{The colour of the palate spline (the default is \code{green}).}
}
\description{
It plots tongue contours from data imported from AAA.
}
