% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_aaa}
\alias{read_aaa}
\title{Read data exported from AAA.}
\usage{
read_aaa(file, column.names, fan.lines = 42, na.rm = FALSE)
}
\arguments{
\item{file}{A character vector.}

\item{column.names}{A character vector with the names of the columns.}

\item{fan.lines}{A number vector with the number of fan lines (the default is \code{42}).}

\item{na.rm}{A boolean string (the default is \code{FALSE}).}
}
\description{
It reads a file with data exported from AAA. The data are automatically
transformed from a wide to a long format (each row has values of X or Y axes
for each fan line). The imported tibble can then be used for plotting and
statistical analysis.
}
