% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_aaa}
\alias{read_aaa}
\title{Read data exported from AAA.}
\usage{
read_aaa(file, column.names, fan.lines = 42, coordinates = "cartesian",
  na.rm = FALSE)
}
\arguments{
\item{file}{The file with AAA data.}

\item{column.names}{The names of the columns withouth including the splines columns.}

\item{fan.lines}{The number of fan lines (the default is \code{42}).}

\item{coordinates}{A string specifying the coordinate system. Possible values are \code{"cartesian"} (the default) and \code{"polar"}.}

\item{na.rm}{Remove NAs (the default is \code{FALSE}).}
}
\description{
It reads a file with data exported from AAA. The data are automatically
transformed from a wide to a long format (each row has values of X or Y axes
for each fan line). The imported tibble can then be used for plotting and
statistical analysis.
}
