% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_aaa}
\alias{read_aaa}
\title{Read tab separated files with AAA spline data.}
\usage{
read_aaa(file, column_names, fan_lines = 42, coordinates = "cartesian",
  na_rm = FALSE)
}
\arguments{
\item{file}{The path of the file with AAA data. It can also be a character vector with multiple paths as separate strings..}

\item{column_names}{The names of the columns withouth including the splines columns.}

\item{fan_lines}{The number of fan lines (the default is \code{42}).}

\item{coordinates}{A string specifying the coordinate system. Possible values are \code{"cartesian"} (the default) and \code{"polar"}.}

\item{na_rm}{Remove NAs (the default is \code{FALSE}).}
}
\description{
It reads a file or a list of files with data exported from AAA. The data are
automatically transformed from a wide to a long format (each row has values
of X or Y axes for each fan line). The imported tibble can then be used for
plotting and statistical analysis.
}
