% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_aaa_data}
\alias{read_aaa_data}
\title{Data import function.}
\usage{
read_aaa_data(file, column_names, fan_lines, coordinates, na_rm)
}
\arguments{
\item{file}{The file or files with AAA data.}

\item{column_names}{The names of the columns withouth including the splines columns.}

\item{fan_lines}{The number of fan lines.}

\item{coordinates}{A string specifying the coordinate system. Possible values are \code{"cartesian"} and \code{"polar"}.}

\item{na_rm}{Remove NAs.}
}
\description{
It reads a file with data exported from AAA. The data are automatically
transformed from a wide to a long format (each row has values of X or Y axes
for each fan line). The imported tibble can then be used for plotting and
statistical analysis.
}
\keyword{internal}
