% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticulate-package.R
\docType{package}
\name{rticulate-package}
\alias{rticulate}
\alias{rticulate-package}
\title{rticulate: Ultrasound Tongue Imaging}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool for processing Articulate Assistant Advanced™ (AAA) export files and plot tongue contour data from any system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/rticulate}
  \item \url{https://stefanocoretta.github.io/rticulate/}
  \item Report bugs at \url{https://github.com/stefanocoretta/rticulate/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com}

}
\keyword{internal}
