% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rticulate.R
\docType{package}
\name{rticulate}
\alias{rticulate}
\alias{rticulate-package}
\title{rticulate: Ultrasound Tongue Imaging in R.}
\description{
This package provides function for processing Articulate Assistant Advanced
(AAA) export files and plot tongue contour data from any system.
}
