% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_coord}
\alias{transform_coord}
\title{Transform the coordinates of spline data}
\usage{
transform_coord(data, to = "polar", origin = NULL)
}
\arguments{
\item{data}{A data set containing the spline coordinates.}

\item{to}{Which system to convert to, as a string, either \code{"polar"} or \code{"cartesian"} (the default is \code{"polar"}).}

\item{origin}{The coordinates of the origin as a vector of \code{c(x, y)} coordinates.}
}
\description{
It transforms the coordinates of spline data between the cartesian and polar
coordinate systems.
}
