% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_coord}
\alias{transform_coord}
\title{Transform the coordinates of spline data}
\usage{
transform_coord(data, to = "polar", origin = NULL, fan_lines = c(10, 25))
}
\arguments{
\item{data}{A data set containing the spline coordinates (cartesian coordinates must be in columns named \code{X} and \code{Y}, polar coordinates in columns named \code{theta} and \code{radius}; these are the defaults in data imported with \code{read_aaa()}).}

\item{to}{Which system to convert to, as a string, either \code{"polar"} or \code{"cartesian"} (the default is \code{"polar"}).}

\item{origin}{The coordinates of the origin as a vector of \code{c(x, y)} coordinates.}

\item{fan_lines}{A numberic vector with two fan lines (the default is \code{c(10, 25)}).}
}
\description{
It transforms the coordinates of spline data between the cartesian and polar
coordinate systems.
}
